requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
mlx>=0.15.0
mlx-vlm>=0.0.8
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
openai-whisper>=20231117
pyttsx3>=2.90
webrtcvad>=2.0.10
pyaudio>=0.2.11
soundfile>=0.12.1
librosa>=0.10.1
requests>=2.31.0
python-dotenv>=1.0.0
